Move sxbp_check() and sxbp_success() to utils.c
